requirements.txt
gradio
matplotlib
torch
opencv-python
easyocr
Pillow>=7.1 
pycocotools>=2.0.2
termcolor>=1.1
yacs>=0.1.8
tabulate
cloudpickle
tqdm>4.29.0
# tensorboardfvcore>=0.1.5<0.1.6 
iopath>=0.1.7<0.1.10
dataclasses; python_version<'3.7'
omegaconf>=2.1<2.4
hydra-core>=1.1
black
packaging
fairscale
timm
scipy>1.5.1
shapely
pygments>=2.2
psutil
panopticapi @ https://github.com/cocodataset/panopticapi/archive/master.zip
git+https://github.com/facebookresearch/fvcore.git
# llama2-wrapper
# --use-pep517
transformers
accelerate
langchain-openai
keras
tensorflow[and-cuda]
tensorflow
Numpy==1.23.5
paddlepaddle-gpu
paddlepaddle
paddleocr>=2.0.1
PyMuPDF==1.21.1